+++ packages/frameworks/react/src/dialog/dialog-description.tsx
@@ -3,14 +3,14 @@ import { forwardRef } from 'react'
 import { ark, type HTMLArkProps } from '../factory'
 import { useDialogContext } from './dialog-context'
 
-export interface DialogDescriptionProps extends HTMLArkProps<'p'> {}
+export interface DialogDescriptionProps extends HTMLArkProps<'div'> {}
 
 export const DialogDescription = forwardRef<HTMLParagraphElement, DialogDescriptionProps>(
   (props, ref) => {
     const api = useDialogContext()
     const mergedProps = mergeProps(api.descriptionProps, props)
 
-    return <ark.p {...mergedProps} ref={ref} />
+    return <ark.div {...mergedProps} ref={ref} />
   },
 )