+++ packages/frameworks/react/src/popover/popover-description.tsx
@@ -3,14 +3,14 @@ import { forwardRef } from 'react'
 import { ark, type HTMLArkProps } from '../factory'
 import { usePopoverContext } from './popover-context'
 
-export interface PopoverDescriptionProps extends HTMLArkProps<'p'> {}
+export interface PopoverDescriptionProps extends HTMLArkProps<'div'> {}
 
 export const PopoverDescription = forwardRef<HTMLParagraphElement, PopoverDescriptionProps>(
   (props, ref) => {
     const api = usePopoverContext()
     const mergedProps = mergeProps(api.descriptionProps, props)
 
-    return <ark.p {...mergedProps} ref={ref} />
+    return <ark.div {...mergedProps} ref={ref} />
   },
 )