+++ packages/frameworks/solid/src/dialog/dialog-description.tsx
@@ -2,11 +2,11 @@ import { mergeProps } from '@zag-js/solid'
 import { ark, type ArkComponent, type HTMLArkProps } from '../factory'
 import { useDialogContext } from './dialog-context'
 
-export interface DialogDescriptionProps extends HTMLArkProps<'p'> {}
+export interface DialogDescriptionProps extends HTMLArkProps<'div'> {}
 
-export const DialogDescription: ArkComponent<'p'> = (props: DialogDescriptionProps) => {
+export const DialogDescription: ArkComponent<'div'> = (props: DialogDescriptionProps) => {
   const dialog = useDialogContext()
   const mergedProps = mergeProps(() => dialog().descriptionProps, props)
 
-  return <ark.p {...mergedProps} />
+  return <ark.div {...mergedProps} />
 }