+++ packages/frameworks/solid/src/popover/popover-description.tsx
@@ -2,11 +2,11 @@ import { mergeProps } from '@zag-js/solid'
 import { ark, type ArkComponent, type HTMLArkProps } from '../factory'
 import { usePopoverContext } from './popover-context'
 
-export interface PopoverDescriptionProps extends HTMLArkProps<'p'> {}
+export interface PopoverDescriptionProps extends HTMLArkProps<'div'> {}
 
-export const PopoverDescription: ArkComponent<'p'> = (props: PopoverDescriptionProps) => {
+export const PopoverDescription: ArkComponent<'div'> = (props: PopoverDescriptionProps) => {
   const api = usePopoverContext()
   const mergedProps = mergeProps(() => api().descriptionProps, props)
 
-  return <ark.p {...mergedProps} />
+  return <ark.div {...mergedProps} />
 }