+++ packages/frameworks/vue/src/dialog/dialog-description.tsx
@@ -2,7 +2,7 @@ import { defineComponent } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
 import { useDialogContext } from './dialog-context'
 
-export interface DialogDescriptionProps extends HTMLArkProps<'p'> {}
+export interface DialogDescriptionProps extends HTMLArkProps<'div'> {}
 
 export const DialogDescription = defineComponent<DialogDescriptionProps>(
   (_, { slots, attrs }) => {